backend/requirements.txt
accelerate==1.8.1
acres==0.2.0
acvl_utils==0.2.5
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
altair==5.5.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.11.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
art==6.4
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.17.0
batchgenerators==0.25.1
batchgeneratorsv2==0.3.0
bcrypt==3.2.2
beautifulsoup4==4.13.3
bitsandbytes==0.46.1
bleach==6.2.0
blinker==1.9.0
blosc2==3.7.2
Brotli==1.1.0
cachetools==5.5.0
certifi==2025.1.31
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.3.0
colorama==0.4.6
comm==0.2.2
comtypes==1.4.11
configobj==5.0.9
configparser==7.1.0
connected-components-3d==3.24.0
contourpy==1.3.1
cryptography==46.0.2
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.12
decorator==5.1.1
deep-translator==1.11.4
defusedxml==0.7.1
dill==0.3.8
distlib==0.3.9
distro==1.9.0
Django==5.2.3
django-cors-headers==4.7.0
dnspython==2.8.0
docopt==0.6.2
docstring_parser==0.16
docx2pdf==0.1.8
dynamic_network_architectures==0.4.2
ecdsa==0.19.1
einops==0.8.1
email_validator==2.2.0
et_xmlfile==2.0.0
executing==2.2.0
fastapi==0.118.0
fastapi-cli==0.0.7
fastjsonschema==2.21.1
feedparser==6.0.11
fft-conv-pytorch==1.2.0
filelock==3.17.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
fonttools==4.56.0
fpdf==1.7.2
fqdn==1.5.1
frozendict==2.4.6
frozenlist==1.5.0
fsspec==2025.2.0
future==1.0.0
g4f==0.4.7.3
generativeai==0.0.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.4.0
google-api-core==2.24.1
google-api-python-client==2.170.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-aiplatform==1.105.0
google-cloud-bigquery==3.35.1
google-cloud-core==2.4.3
google-cloud-resource-manager==1.14.2
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.26.0
google-generativeai==0.4.1
google-resumable-media==2.7.2
googleapis-common-protos==1.69.0
googlemaps==4.10.0
graphviz==0.21
greenlet==3.1.1
grpc-google-iam-v1==0.14.2
grpcio==1.71.0rc2
grpcio-status==1.63.0rc1
gspread==6.1.4
h11==0.16.0
hf-xet==1.1.5
html5lib==1.1
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
imagecodecs==2025.8.2
imageio==2.37.0
ipykernel==6.29.5
ipython==8.32.0
isodate==0.6.1
isoduration==20.11.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jose==1.0.0
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.15.2
linecache2==1.0.0
looseversion==1.3.0
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdurl==0.1.2
mistune==3.1.2
motor==3.7.1
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multiprocess==0.70.16
multitasking==0.0.11
mysql-connector-python==9.2.0
mysqlclient==2.2.7
narwhals==1.22.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ndindex==1.10.0
nest-asyncio==1.6.0
networkx==3.4.2
nibabel==5.3.2
ninja==1.11.1.3
nltk==3.9.1
nnunetv2==2.6.2
notebook_shim==0.2.4
numexpr==2.11.0
numpy==2.1.1
oauthlib==3.2.2
openai==1.98.0
opencv-python==4.11.0.86
openpyxl==3.1.5
orjson==3.10.18
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
passlib==1.7.4
pdf2image==1.17.0
pdfkit==1.0.0
pdfminer.six==20231228
pdfplumber==0.11.5
peewee==3.17.9
pefile==2023.2.7
peft==0.16.0
pillow==11.1.0
platformdirs==4.3.6
playwright==1.53.0
primp==0.12.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.2.1
proto-plus==1.26.0
protobuf==4.25.8
psutil==7.0.0
psycopg2-binary==2.9.9
pure_eval==0.2.3
puremagic==1.28
py-cpuinfo==9.0.0
pyarrow==19.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAutoGUI==0.9.54
pyclipper==1.3.0.post6
pycparser==2.23
pycryptodome==3.21.0
pydantic==2.11.10
pydantic-extra-types==2.10.5
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydeck==0.9.1
pydot==3.0.4
pyee==13.0.0
pygame==2.6.1
PyGetWindow==0.0.9
Pygments==2.19.1
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyJWT==2.1.0
pymongo==4.15.2
PyMsgBox==1.0.9
pyparsing==3.2.1
PyPDF2==3.0.1
pypdfium2==4.30.1
pyperclip==1.9.0
pypiwin32==223
PyRect==0.2.0
PyScreeze==1.0.1
PySocks==1.7.1
pytesseract==0.3.13
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.0.0
python-jose==3.5.0
python-json-logger==3.2.1
python-multipart==0.0.20
python-pptx==1.0.2
pytorch-lightning==2.5.5
pyttsx3==2.99
pytweening==1.2.0
pytz==2025.1
pywhatkit==5.4
pywin32==308
pywin32-ctypes==0.2.3
pywinpty==2.0.15
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.1
regex==2024.11.6
reportlab==4.3.0
requests==2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rich-toolkit==0.14.7
rpds-py==0.22.3
rsa==4.9.1
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.0
scipy==1.15.0
seaborn==0.13.2
selenium==4.29.0
Send2Trash==1.8.3
setuptools==80.9.0
sgmllib3k==1.0.0
shapely==2.1.1
shellingham==1.5.4
SimpleITK==2.3.1
simplejson==3.19.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.6
SQLAlchemy==2.0.41
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
streamlit==1.41.1
sympy==1.13.3
tabulate==0.9.0
tenacity==8.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.6.11
timm==1.0.19
tinycss2==1.4.0
tokenizers==0.21.2
toml==0.10.2
tomli==2.2.1
torch==2.7.1+cpu
torchaudio==2.7.1+cpu
torchmetrics==1.8.2
torchvision==0.22.1+cpu
tornado==6.4.2
tqdm==4.67.1
traceback2==1.4.0
traitlets==5.14.3
traits==7.0.2
transformers==4.53.1
trio==0.29.0
trio-websocket==0.12.2
trl==0.19.0
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
uber-rides==0.6.0
ujson==5.10.0
unittest2==1.1.0
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.37.0
virtualenv==20.28.0
virtualenvwrapper-win==1.2.7
watchdog==6.0.0
watchfiles==1.1.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
wikipedia==1.4.0
wsproto==1.2.0
xgboost==3.0.0
xlsxwriter==3.2.9
xxhash==3.5.0
yacs==0.1.8
yarl==1.18.3
yfinance==0.2.54
